Remove analysis.read_prep subpackage
  * Move FastQC code in this subpackage to the analysis.fastqc module (one function)
* Move the skewer module in the subpackage to analysis.skewer
